Tests/LibWeb/Text/input/DOM/HTMLCollection-empty-string-as-property-name.html

<script src="../include.js"></script>
<div id></div>
<div id=haxx></div>
<script>
    test(() => {
        var c = document.getElementsByTagName("*");
        println("Empty string in collection: " + ("" in c));
        println("'haxx' in collection: " + ("haxx" in c));
    });
</script>
