Tests/LibWeb/Text/expected/DOM/HTMLCollection-empty-string-as-property-name.txt

Empty string in collection: false
'haxx' in collection: true
